+++ src/components/common/PhraseCRUD/PhraseCRUD.module.css
@@ -1,0 +1,11 @@
+.farsiVariation {
+  position: relative;
+}
+
+.farsiVariationRemove {
+  position: absolute;
+  top: 9px;
+  right: -5px;
+  background: #fff;
+  border-radius: 50%;
+}
